requirements.txt
einops==0.8.0
kornia==0.6.5
mmselfsup==1.0.0
natsort==8.1.0
numpy==1.22.4
opencv_python==4.5.4.60
pandas==2.2.2
Pillow==9.5.0
Pillow==10.3.0
pytorch_wavelets==1.3.0
scikit_learn==1.4.1.post1
scipy==1.7.3
timm==0.3.2
torch==1.11.0+cu113
torch_geometric==2.5.3
torchvision==0.12.0
tqdm==4.65.2
wandb==0.12.11
warmup_scheduler==0.3
